Missing renames of mail confirmation Text changes proposed by @Twometer
