+++ src/SkynetServer.Web/Views/MailConfirmation/Invalid.cshtml
@@ -5,4 +5,4 @@
 
 <h2 class="main-header">Invalid confirmation code</h2>
 
-<p>The confirmation code you entered is invalid. After creating your account the confirmation code is valid for 24 hours.</p>
+<p>The confirmation code you entered is invalid. Please note that after creating your account, the confirmation code is only valid for 24 hours.</p>
